DDA-191: Refactor welcome screen
